bugfixes and first script
